List's rows method improved sort
Now appends any rows that are empty on the sortable field to the bottom
of the sort query, regardless of whether it's ASC or DESC.

This should be improved upon, when it is ASC should append to the
BEGINNING of the rows. Need to write a more complex function to order
and concatenate the queries.